% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{cjustin.utils}
\alias{cjustin.utils}
\alias{cjustin.utils-package}
\title{A custom function package}
\description{
This package consists of utility functions that I seem to commonly use to
avoid repeating myself.
}
\details{
This package does not have a unifying theme. Functions range from adding
custom themes to ggplot2 plots to reading and writing tab seperated files. If
all goes well and my usage of R increases, I will likely add more functions.

Functions added:
\code{\link{peek}},
\code{\link{reorderdesc}},
\code{\link{writeTSV}},
\code{\link{readTSV}},
\code{\link{formatFig}},
}

