% Generated by roxygen2 (4.0.2): do not edit by hand
\name{formatFig}
\alias{formatFig}
\title{Format ggplot2 figures with default theme elements}
\usage{
formatFig(type = "bioinformatics")
}
\arguments{
\item{type}{string used to determine what theme to use}
}
\value{
a list containing theme options for ggplot2
}
\description{
Formats ggplot2 figure with themes that I have custom made for
  figures in various journals. As I make more figure I may have more defaults
  to add but currently has the defaults I recently for the journal of
  bioinformatics.
}
\details{
Themes: bioinformatics - a black and white theme with large font
}
\examples{
library(ggplot2)
testDat <- data.frame( treatment = sample( c("PLACEBO", "300 MG", "600 MG", "1200 MG"),
                                         100, replace=TRUE ))
p <- ggplot(testDat, aes(treatment)) + geom_bar()
p <- p + formatFig("bioinformatics")
}
\author{
Justin Chu
}
\seealso{
\code{\link{ggplot2}}
}

