% Generated by roxygen2 (4.0.2): do not edit by hand
\name{peek}
\alias{peek}
\title{Return the part of an Object}
\usage{
peek(x, ...) peek(x, n = 6, ...) peek(x, n = 6, loc = 0.5)
}
\arguments{
\item{x}{A dataframe}

\item{n}{Number of element wanted to be returned}

\item{loc}{The location as proportion of file to draw values from}
}
\description{
Returns parts of a data frame.
}
\details{
Uses \code{loc} to determine location of dataframe to look at. Will
  only return rows that are before the end of the file. Zero or negative
  \code{n} values will cause function to return empty dataframes.
}
\examples{
peek(x)
peek(x, 10)
peek(x, 10, 0.2)
peek(x, loc = 0.8)
}
\author{
Justin Chu
}

