% Generated by roxygen2 (4.0.2): do not edit by hand
\name{peek}
\alias{peek}
\title{Return the part of an Object}
\usage{
peek(x, ...) peek(x, n = 6, ...) peek(x, n = 6, loc = 0.5)
}
\arguments{
\item{x}{A dataframe}

\item{n}{Number of element wanted to be returned}

\item{loc}{The location as proportion of file to draw values from}
}
\value{
a subset of the input dataframe
}
\description{
Returns parts of a data frame.
}
\details{
Uses \code{loc} to determine location of dataframe to look at. Will
  only return rows that are before the end of the file. Zero or negative
  \code{n} values will cause errors to be returned
}
\examples{
testDat <- data.frame( treatment = sample( c("PLACEBO", "300 MG", "600 MG", "1200 MG"),
                                         100, replace=TRUE ))
peek(testDat)
peek(testDat, 10)
peek(testDat, 10, 0.2)
peek(testDat, loc = 0.8)
}
\author{
Justin Chu
}

